Для ускорения перевел все float в double

+++ directProblemDimensionalKvadraturOLd/main.cpp
@@ -6,24 +6,24 @@ using namespace std;
 int main()
 {
 	size_t N = numberPartitionPosize_ts_N;
-	float h = (float)domainInHomogeneity_R / numberPartitionPosize_ts_N;
+	double h = (double)domainInHomogeneity_R / numberPartitionPosize_ts_N;
 
 	// выделение памяти для точного решения xi и акустического поля u
-	float **xi;
-	complex<float> ** u;
-	xi = new float *[N + 1];
-	u = new complex<float> *[N + 1];
+	double **xi;
+	complex<double> ** u;
+	xi = new double *[N + 1];
+	u = new complex<double> *[N + 1];
 	for (size_t i = 0; i <= N; ++i)
 	{
-		xi[i] = new float[N + 1];
-		u[i] = new complex<float>[N + 1];
+		xi[i] = new double[N + 1];
+		u[i] = new complex<double>[N + 1];
 	}
 	// задание точного решения \xi
 	for (size_t i = 0; i <= N; ++i)
 	{
 		for (size_t j = 0; j <= N; ++j)
 		{
-			xi[i][j] = exp(-(i * h - 6.0f) * (i * h - 6.0f) - (j * h - 6.0f) * (j * h - 6.0f));
+			xi[i][j] = exp(-(i * h - 6.0) * (i * h - 6.0) - (j * h - 6.0) * (j * h - 6.0));
 		}
 	}
 
@@ -39,36 +39,36 @@ int main()
 	f_xi.close();
 
 	// выделяем память под основные матрицы
-	complex<float> ****a;
-	a = new complex<float> ***[N + 1];
+	complex<double> ****a;
+	a = new complex<double> ***[N + 1];
 	for (size_t i = 0; i <= N; ++i)
 	{
-		a[i] = new complex<float>**[N + 1];
+		a[i] = new complex<double>**[N + 1];
 		for (size_t j = 0; j <= N; ++j)
 		{
-			a[i][j] = new complex<float>*[N + 1];
+			a[i][j] = new complex<double>*[N + 1];
 			{
 				for (size_t p = 0; p <= N; ++p)
 				{
-					a[i][j][p] = new complex<float>[N + 1];
+					a[i][j][p] = new complex<double>[N + 1];
 				}
 			}
 		}
 	}
-	complex<float> **b;
-	b = new complex<float> *[N + 1];
+	complex<double> **b;
+	b = new complex<double> *[N + 1];
 	for (size_t i = 0; i <= N; ++i)
 	{
-		b[i] = new complex<float>[N + 1];
+		b[i] = new complex<double>[N + 1];
 	}
-	complex<float> ***overline_a;
-	overline_a = new complex<float> **[N + 1];
+	complex<double> ***overline_a;
+	overline_a = new complex<double> **[N + 1];
 	for (size_t i = 0; i <= N; ++i)
 	{
-		overline_a[i] = new complex<float>*[N + 1];
+		overline_a[i] = new complex<double>*[N + 1];
 		for (size_t j = 0; j <= N; ++j)
 		{
-			overline_a[i][j] = new complex<float>[N + 1];
+			overline_a[i][j] = new complex<double>[N + 1];
 		}
 	}
 
@@ -78,21 +78,21 @@ int main()
 	clock_t timeStart, timeFinish;
 	timeStart = clock();
 	// для индексов метода квадратур
-	float *index;
-	index = new float[N + 1];
+	double *index;
+	index = new double[N + 1];
 	for (size_t i = 1; i < N; ++i)
 	{
 		if (i % 2 != 0)
 		{
-			index[i] = (float)4 / 3;
+			index[i] = 4 / 3;
 		}
 		else
 		{
-			index[i] = (float)2 / 3;
+			index[i] = 2 / 3;
 		}
 	}
-	index[0] = (float)1 / 3;
-	index[N] = (float)1 / 3;
+	index[0] = 1 / 3;
+	index[N] = 1 / 3;
 	// нахождение массива a
 	for (size_t i = 0; i <= N; ++i)
 	{
@@ -130,7 +130,7 @@ int main()
 	{
 		for (size_t j = 0; j <= N; ++j)
 		{
-			b[i][j] = { 0.0f, 0.0f };
+			b[i][j] = { 0.0, 0.0 };
 			for (size_t p = 0; p < N; ++p)
 			{
 				for (size_t q = 0; q < N; ++q)
@@ -147,7 +147,7 @@ int main()
 
 	//печатаем время работы
 	timeFinish = clock();
-	float d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	double d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Time calculation of basic matrices " << d << endl;
 	timeStart = clock();
 
@@ -192,7 +192,7 @@ int main()
 	f_b.close();
 	//печатаем время работы
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Download time major arrays " << d << endl;
 	timeStart = clock();
 
@@ -210,7 +210,7 @@ int main()
 	ofstream f_Source_02("f_Source_02.txt");
 	for (size_t j = 0; j <= N; ++j)
 	{
-		f_Source_02 << fixed << setprecision(12) << f_01(11.0f, j * h) << " ";
+		f_Source_02 << fixed << setprecision(12) << f_01(11.0, j * h) << " ";
 	}
 	f_Source_02.close();
 
@@ -227,7 +227,7 @@ int main()
 	ofstream f_Source_04("f_Source_04.txt");
 	for (size_t j = 0; j <= N; ++j)
 	{
-		f_Source_04 << fixed << setprecision(12) << f_02(11.0f, j * h) << " ";
+		f_Source_04 << fixed << setprecision(12) << f_02(11.0, j * h) << " ";
 	}
 	f_Source_04.close();
 
@@ -244,7 +244,7 @@ int main()
 	ofstream f_Source_06("f_Source_06.txt");
 	for (size_t j = 0; j <= N; ++j)
 	{
-		f_Source_06 << fixed << setprecision(12) << f_03(11.0f, j * h) << " ";
+		f_Source_06 << fixed << setprecision(12) << f_03(11.0, j * h) << " ";
 	}
 	f_Source_06.close();
 
@@ -261,7 +261,7 @@ int main()
 	ofstream f_Source_08("f_Source_08.txt");
 	for (size_t j = 0; j <= N; ++j)
 	{
-		f_Source_08 << fixed << setprecision(12) << f_04(11.0f, j * h) << " ";
+		f_Source_08 << fixed << setprecision(12) << f_04(11.0, j * h) << " ";
 	}
 	f_Source_08.close();
 
@@ -278,40 +278,40 @@ int main()
 	ofstream f_Source_10("f_Source_10.txt");
 	for (size_t j = 0; j <= N; ++j)
 	{
-		f_Source_10 << fixed << setprecision(12) << f_05(11.0f, j * h) << " ";
+		f_Source_10 << fixed << setprecision(12) << f_05(11.0, j * h) << " ";
 	}
 	f_Source_10.close();
 
 	//печатаем время работы
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "The computation time of the source function " << d << endl;
 	timeStart = clock();
 
 	// для нахождения u^(1) составляем СЛАУ основная матрица * u^(1) = правой части
 	// substantiveMatrix[ii][jj] * numbered_u[jj] = rightPartEequation[ii]
 	size_t N_squared = (N + 1) * (N + 1);
-	complex<float> *numbered_u, *rightPartEquation;
-	rightPartEquation = new complex<float>[N_squared];
-	numbered_u = new complex<float>[N_squared];
+	complex<double> *numbered_u, *rightPartEquation;
+	rightPartEquation = new complex<double>[N_squared];
+	numbered_u = new complex<double>[N_squared];
 	//память под основную матрицу
-	complex<float> **substantiveMatrix;
-	substantiveMatrix = new complex<float> *[N_squared];
+	complex<double> **substantiveMatrix;
+	substantiveMatrix = new complex<double> *[N_squared];
 	for (size_t i = 0; i < N_squared; ++i)
 	{
-		substantiveMatrix[i] = new complex<float>[N_squared];
+		substantiveMatrix[i] = new complex<double>[N_squared];
 	}
 	for (size_t i = 0; i < N_squared; ++i)
 	{
 		for (size_t j = 0; j < N_squared; ++j)
 		{
-			substantiveMatrix[i][j] = { 0.0f, 0.0f };
+			substantiveMatrix[i][j] = { 0.0, 0.0 };
 		}
 	}
 
 	// выделение памяти для overline_u_0
-	complex<float> *overline_u;
-	overline_u = new complex<float>[N + 1];
+	complex<double> *overline_u;
+	overline_u = new complex<double>[N + 1];
 	//
 	//счет основной матрицы
 	//
@@ -322,7 +322,7 @@ int main()
 		for (size_t j = 0; j <= N; ++j)
 		{
 			ii = i * (N + 1) + j;
-			complex<float> sumOfTheCoefficients = 0.0f;
+			complex<double> sumOfTheCoefficients = 0.0;
 			for (size_t p = 0; p < N; ++p)
 			{
 				for (size_t q = 0; q < N; ++q)
@@ -335,13 +335,13 @@ int main()
 					}
 				}
 			}
-			substantiveMatrix[ii][ii] += 1.0f;
+			substantiveMatrix[ii][ii] += 1.0;
 			substantiveMatrix[ii][ii] -= sumOfTheCoefficients * xi[i][j];
 			substantiveMatrix[ii][ii] += b[i][j] * xi[i][j];
 		}
 	}
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "The computation time of the matrix inside the squared " << d << endl;
 	timeStart = clock();
 
@@ -370,7 +370,7 @@ int main()
 		u[coordinate_x][coordinate_y] = numbered_u[l];
 	}
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in R for 1 source " << d << endl;
 	timeStart = clock();
 	//
@@ -378,7 +378,7 @@ int main()
 	//
 	for (size_t i = 0; i <= N; ++i)
 	{
-		overline_u[i] = f_01(11.0f, i * h);
+		overline_u[i] = f_01(11.0, i * h);
 	}
 	for (size_t j = 0; j <= N; ++j)
 	{
@@ -398,7 +398,7 @@ int main()
 	}
 	f_overline_u_1.close();
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in X for 1 source " << d << endl;
 	timeStart = clock();
 
@@ -427,7 +427,7 @@ int main()
 		u[coordinate_x][coordinate_y] = numbered_u[l];
 	}
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in R for 2 source " << d << endl;
 	timeStart = clock();
 	//
@@ -435,7 +435,7 @@ int main()
 	//
 	for (size_t i = 0; i <= N; ++i)
 	{
-		overline_u[i] = f_02(11.0f, i * h);
+		overline_u[i] = f_02(11.0, i * h);
 	}
 	for (size_t j = 0; j <= N; ++j)
 	{
@@ -455,7 +455,7 @@ int main()
 	}
 	f_overline_u_2.close();
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in X for 2 source " << d << endl;
 	timeStart = clock();
 
@@ -483,7 +483,7 @@ int main()
 		u[coordinate_x][coordinate_y] = numbered_u[l];
 	}
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in R for 3 source " << d << endl;
 	timeStart = clock();
 	//
@@ -491,7 +491,7 @@ int main()
 	//
 	for (size_t i = 0; i <= N; ++i)
 	{
-		overline_u[i] = f_03(11.0f, i * h);
+		overline_u[i] = f_03(11.0, i * h);
 	}
 	for (size_t j = 0; j <= N; ++j)
 	{
@@ -511,7 +511,7 @@ int main()
 	}
 	f_overline_u_3.close();
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in X for 3 source " << d << endl;
 	timeStart = clock();
 
@@ -539,7 +539,7 @@ int main()
 		u[coordinate_x][coordinate_y] = numbered_u[l];
 	}
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in R for 4 source " << d << endl;
 	timeStart = clock();
 	//
@@ -547,7 +547,7 @@ int main()
 	//
 	for (size_t i = 0; i <= N; ++i)
 	{
-		overline_u[i] = f_04(11.0f, i * h);
+		overline_u[i] = f_04(11.0, i * h);
 	}
 	for (size_t j = 0; j <= N; ++j)
 	{
@@ -567,7 +567,7 @@ int main()
 	}
 	f_overline_u_4.close();
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in X for 4 source " << d << endl;
 	timeStart = clock();
 
@@ -595,7 +595,7 @@ int main()
 		u[coordinate_x][coordinate_y] = numbered_u[l];
 	}
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in R for 5 source " << d << endl;
 	timeStart = clock();
 	//
@@ -603,7 +603,7 @@ int main()
 	//
 	for (size_t i = 0; i <= N; ++i)
 	{
-		overline_u[i] = f_05(11.0f, i * h);
+		overline_u[i] = f_05(11.0, i * h);
 	}
 	for (size_t j = 0; j <= N; ++j)
 	{
@@ -623,7 +623,7 @@ int main()
 	}
 	f_overline_u_5.close();
 	timeFinish = clock();
-	d = (float)(timeFinish - timeStart) / CLOCKS_PER_SEC;
+	d = (double)(timeFinish - timeStart) / CLOCKS_PER_SEC;
 	cout << "Finding the acoustic pressure in X for 5 source " << d << endl;
 	timeStart = clock();
